Assignment4/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
mpmath==1.3.0
numpy==2.2.4
pycryptodome==3.22.0
requests==2.32.3
scipy==1.15.2
sympy==1.13.3
urllib3==2.3.0
python-secrets
